Redesigned the API for the library
Completely redesigned the API. Instead of having functions there is now
the FancyPrinter struct, which can be instantiated and used to print text
by calling the print() method. This struct can be constructed either by
calling the builder() method, or the FancyPrinterBuilder struct. This
uses the builder design pattern.

Also added an additional animation field in the FancyPrinter struct that
allows the user to determine what animation they would like to use.
Currently there is the Character Cycling one that the program originally
used, and currently there is also the Typing animation, which looks
exactly how you would imagine it to.

I think I am going to try and make this a cli tool for shell scripts,
but also a library for rust cli applications too.