Refactored lib.rs to not use any Clap components
Added an intermediate enum called AnimationOptions that derives the
needed clap functions, and converts itself to the Animation enum so that
the lib package did not use any clap code. I'm not sure how it works
completely, but I'm hoping that when importing just the library clap is
not added as a dependency.